requirements.txt
fastapi
uvicorn
websockets
twilio
python-dotenv
scipy
numpy
